+++ extensions/material-helper/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.9
 
+- feat: change auto fill code to React Generator extension. 
 - fix: the code in the editor is incomplete will cause the AST parser error.
 
 ## 1.0.8